+++ catalog/src/main/java/com/codecool/catalog/repository/ItemRepository.java
@@ -1,6 +1,6 @@
 package com.codecool.catalog.repository;
 
-import com.codecool.catalog.model.Item;
+import com.codecool.catalog.modell.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
